requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
python-dotenv==0.20.0
Flask-Ngrok==0.0.25
requests==2.32.3
urllib3==2.0.7
line-bot-sdk==3.13.0